+++ R/scto_write.R
@@ -24,7 +24,8 @@
 #' r = scto_write(auth, data, 'my_dataset', 'My Dataset')
 #' }
 #'
-#' @seealso [scto_auth()], [scto_read()], [scto_get_attachments()]
+#' @seealso [scto_auth()], [scto_meta()], [scto_read()],
+#'   [scto_get_attachments()]
 #'
 #' @export
 scto_write = function(